DDP-5713 | added custom tabs to overcome slow performance, many init of component
