+++ src/app/participant-page/participant-page.component.ts
@@ -30,6 +30,8 @@ import {PatchUtil} from "../utils/patch.model";
 import { ParticipantUpdateResultDialogComponent } from "../dialogs/participant-update-result-dialog.component";
 import { AddFamilyMemberComponent } from "../popups/add-family-member/add-family-member.component";
 import { Sample } from "../participant-list/models/sample.model";
+import { Tab } from "../tabs/tab.model";
+import { TabComponent } from "../tabs/tab.component";
 
 var fileSaver = require( "file-saver/FileSaver.js" );
 
@@ -109,6 +111,8 @@ export class ParticipantPageComponent implements OnInit, OnDestroy {
 
   private payload = {};
 
+  participantTabs: TabComponent[] = [];
+
   constructor( private auth: Auth, private compService: ComponentService, private dsmService: DSMService, private router: Router,
                private role: RoleService, private util: Utils, private route: ActivatedRoute, public dialog: MdDialog) {
     if (!auth.authenticated()) {
@@ -159,6 +163,10 @@ export class ParticipantPageComponent implements OnInit, OnDestroy {
 
     clearInterval(this.checkParticipantStatusInterval);
 
+  }
+
+  putTab(tab: TabComponent) {
+    this.participantTabs.push(tab);
   }
 
   showFamilyMemberPopUpOnClick() {
@@ -1100,24 +1108,7 @@ export class ParticipantPageComponent implements OnInit, OnDestroy {
     );
   }
 
-  getParticipantData(fieldSetting: FieldSettings, relative: ParticipantData) {
-    if (this.participant != null && this.participant.participantData != null && relative.dataId != null && fieldSetting.columnName != null) {
-      let kitFieldsDict = {'DATE_KIT_RECEIVED': 'receiveDate', 'DATE_KIT_SENT': 'scanDate', 'KIT_TYPE_TO_REQUEST': 'kitType'};
-      if (fieldSetting.displayType === 'SAMPLE') {
-        let sample: Sample = this.participant.kits.find(kit => kit.bspCollaboratorSampleId === relative.data[fieldSetting.columnName]);
-        if (sample && kitFieldsDict[fieldSetting.columnName] && sample[kitFieldsDict[fieldSetting.columnName]]) {
-          return sample[kitFieldsDict[fieldSetting.columnName]];
-        } else {
-          return "";
-        }
-      }
-      let participantData = this.participant.participantData.find(participantData => participantData.dataId === relative.dataId);
-      if (participantData != null && participantData.data != null && participantData.data[fieldSetting.columnName] != null) {
-        return participantData.data[fieldSetting.columnName];
-      }
-    }
-    return "";
-  }
+
 
   getDisplayName(displayName: string, columnName: string) {
     if (displayName.indexOf('#') > -1) {
@@ -1142,213 +1133,15 @@ export class ParticipantPageComponent implements OnInit, OnDestroy {
     }
   }
 
-  getActivityData(fieldSetting: FieldSettings) {
-    //type was activity or activity_staff and no saved staff answer. therefore lookup the activity answer
-    if (fieldSetting != null && fieldSetting.possibleValues != null && fieldSetting.possibleValues[0] != null && fieldSetting.possibleValues[0].value != null) {
-      let tmp: string[] = fieldSetting.possibleValues[ 0 ].value.split( '.' );
-      if (tmp != null && tmp.length > 1) {
-        if (tmp[ 0 ] === 'profile') {
-          return this.participant.data.profile[tmp[1]];
-        }
-        else {
-          if (this.participant != null && this.participant.data != null && this.participant.data.activities != null) {
-            let activity: ActivityData = this.participant.data.activities.find( activity => activity.activityCode === tmp[ 0 ] );
-            if (activity != null && activity.questionsAnswers != null) {
-              let questionAnswer = activity.questionsAnswers.find( questionAnswer => questionAnswer.stableId === tmp[ 1 ] );
-              if (questionAnswer != null) {
-                if (tmp.length == 2) {
-                  if (typeof questionAnswer.answer === "boolean") {
-                    if (questionAnswer.answer) {
-                      return "Yes";
-                    }
-                    return "No";
-                  }
-                  if (questionAnswer.answer instanceof Array) {
-                    return questionAnswer.answer[0];
-                  }
-                  return questionAnswer.answer;
-                }
-                else if (tmp.length === 3) {
-                  if (fieldSetting.possibleValues != null && fieldSetting.possibleValues[ 0 ] != null && fieldSetting.possibleValues[ 0 ].type != null && fieldSetting.possibleValues[ 0 ].type === "RADIO") {
-                    if (questionAnswer.answer != null) {
-                      let found = questionAnswer.answer.find( answer => answer === tmp[ 2 ] )
-                      if (found != null) {
-                        return "Yes";
-                      }
-                      return "No";
-                    }
-                  }
-                  else if (this.activityDefinitions != null) {
-                    let definition: ActivityDefinition = this.activityDefinitions.find( definition => definition.activityCode === tmp[ 0 ] );
-                    if (definition != null && definition.questions != null) {
-                      let question = definition.questions.find( question => question.stableId === tmp[ 1 ] );
-                      if (question != null && question.childQuestions != null) {
-                        for (let i = 0; i < question.childQuestions.length; i++) {
-                          if (question.childQuestions[ i ] != null && question.childQuestions[ i ].stableId === tmp[ 2 ] && questionAnswer.answer[ 0 ][ i ] != null) {
-                            return questionAnswer.answer[ 0 ][ i ];
-                          }
-                        }
-                      }
-                      else if (question != null && question.options != null) {
-
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    return "";
-  }
+  
 
   getTabsSettingList() {
     return this.settings['null'];
   }
 
-  getActivityOptions(fieldSetting: FieldSettings) {
-    if (fieldSetting.displayType === 'ACTIVITY' || fieldSetting.displayType === 'ACTIVITY_STAFF') {
-      if (fieldSetting.possibleValues != null && fieldSetting.possibleValues[0] != null && fieldSetting.possibleValues[0].value != null
-        && fieldSetting.possibleValues[0].type != null) {
-        let tmp: string[] = fieldSetting.possibleValues[ 0 ].value.split( '.' );
-        if (tmp != null && tmp.length > 1) {
-          if (tmp.length == 2) {
-            if (this.activityDefinitions != null) {
-              let definition: ActivityDefinition = this.activityDefinitions.find( definition => definition.activityCode === tmp[ 0 ] );
-              if (definition != null && definition.questions != null) {
-                let question = definition.questions.find( question => question.stableId === tmp[ 1 ] );
-                if (question != null) {// && question.options != null) {
-                  if (question.questionType !== 'BOOLEAN' && question.options != null) {
-                    let options: NameValue[] = [];
-                    for (let i = 0; i < question.options.length; i++) {
-                      options.push( new NameValue( question.options[ i ].optionText, question.options[ i ].optionStableId ));
-                    }
-                    return options;
-                  }
-                  else {
-                    let options: string[] = [];
-                    options.push( "Yes" );
-                    options.push( "No" );
-                    return options;
-                  }
-                }
-              }
-            }
-          }
-          else if (tmp.length === 3) {
-            let options: string[] = [];
-            options.push( "Yes" );
-            options.push( "No" );
-            return options;
-          }
-        }
-      }
-    }
-    return [];
-  }
+  
 
-  formPatch(value: any, fieldSetting: FieldSettings, groupSetting: FieldSettings, dataId?: string) {
-    if (value.value && value.dataId && value.fieldSetting && value.groupSetting) {
-      let incomingValue = value;
-      value = incomingValue.value;
-      fieldSetting = incomingValue.fieldSetting;
-      groupSetting = incomingValue.groupSetting;
-      dataId = incomingValue.dataId;
-    }
-    debugger;
-    if (fieldSetting == null || fieldSetting.fieldType == null) {
-      this.errorMessage = "Didn't save change";
-      return;
-    }
-    let fieldTypeId = fieldSetting.fieldType;
-    if (groupSetting != null) {
-      fieldTypeId = groupSetting.fieldType;
-    }
-    if (this.participant != null && this.participant.participantData != null && fieldTypeId != null && fieldSetting.columnName != null && dataId != null) {
-      let participantData: ParticipantData = this.participant.participantData.find(participantData => participantData.dataId === dataId);
-      if (participantData == null) {
-        let data: { [ k: string ]: any } = {};
-        data[fieldSetting.columnName] = value;
-        participantData = new ParticipantData (null, fieldTypeId, data );
-        this.participant.participantData.push(participantData);
-      }
-      if (participantData != null && participantData.data != null) {
-        participantData.data[fieldSetting.columnName] = value;
-
-        let nameValue: { name: string, value: any }[] = [];
-        nameValue.push({name: "d.data", value: JSON.stringify(participantData.data)});
-        let participantDataSec: ParticipantData = null;
-        let actionPatch: Value[] = null;
-        if (fieldSetting.actions != null) {
-          fieldSetting.actions.forEach(( action ) => {
-            if (action != null && action.name != null && action.name != undefined && action.type != null && action.type != undefined) {
-              participantDataSec = this.participant.participantData.find(participantData => participantData.fieldTypeId === action.type);
-              if (participantDataSec == null) {
-                if (action.type !== 'ELASTIC_EXPORT') {
-                  let data: { [ k: string ]: any } = {};
-                  data[ action.name ] = action.value;
-                  participantDataSec = new ParticipantData( null, action.type, data );
-                }
-                else {
-                  if (actionPatch === null) {
-                    actionPatch = [];
-                  }
-                  actionPatch.push(action);
-                }
-              }
-              if (participantDataSec != null && participantDataSec.data != null) {
-                participantDataSec.data[ action.name ] = action.value;
-                nameValue.unshift({name: "d.data", value: JSON.stringify(participantDataSec.data)});
-              }
-            }
-          });
-        }
-        if (fieldSetting.fieldType === "RADIO" && fieldSetting.possibleValues != null) {
-          let possibleValues = fieldSetting.possibleValues;
-          let possibleValue = possibleValues.find(value => value.name === fieldSetting.columnName && value.values != null)
-        }
-
-        let participantId = this.participant.data.profile[ "guid" ];
-        if (this.participant.data.profile[ "legacyAltPid" ] != null && this.participant.data.profile[ "legacyAltPid" ] != undefined && this.participant.data.profile[ "legacyAltPid" ] !== '') {
-          participantId = this.participant.data.profile[ "legacyAltPid" ];
-        }
-        let patch = {
-          id: participantData.dataId,
-          parent: "participantDataId",
-          parentId: participantId,
-          user: this.role.userMail(),
-          fieldId: fieldTypeId,
-          realm:  localStorage.getItem( ComponentService.MENU_SELECTED_REALM ),
-          nameValues: nameValue,
-          actions: actionPatch,
-        };
-
-        this.dsmService.patchParticipantRecord( JSON.stringify( patch ) ).subscribe(// need to subscribe, otherwise it will not send!
-          data => {
-            let result = Result.parse( data );
-            if (result.code === 200) {
-              if (result.body != null && result.body !== "") {
-                let jsonData: any | any[] = JSON.parse( result.body );
-                if (jsonData.participantDataId !== undefined && jsonData.participantDataId !== "") {
-                  if (participantData != null) {
-                    participantData.dataId = jsonData.participantDataId;
-                  }
-                }
-              }
-            }
-            this.patchFinished = true;
-          },
-          err => {
-            if (err._body === Auth.AUTHENTICATION_ERROR) {
-              this.router.navigate( [ Statics.HOME_URL ] );
-            }
-          }
-        );
-      }
-    }
-  }
+  
 
   createRelativeTabHeading(data: any): string {
     if (data) {
