+++ src/app/tabs/tab.model.ts
@@ -1,0 +1,6 @@
+export class Tab {
+    constructor(public title: string, public active: boolean) {
+        this.title = title;
+        this.active = active;
+    }
+}
