+++ src/app/tabs/tabs.component.html
@@ -1,0 +1,3 @@
+<p>
+  tabs works!
+</p>
